requirements.txt
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-core==1.7.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
rsa==4.7.2
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
urllib3==1.26.5
w3lib==1.22.0
zope.interface==5.4.0
